Fix bg_flag argument to addjob function and add jobs to builtin_cmd
Also moved the message printed out upon starting a background prcess
in the block which runs with SIGCHLD blocked. This is done to ensure that
job is not terminated and deleted before the message can be desplayed.
This is important because pid2jid must be able to find job in the list
to print the message