Modify eval, builtin_cmd, do_bgfg to use common pid across all functions
This was done to ensure that waitfg function is only called once in eval.
If the shell command is bgfg then do_bgfg function writes the corresponding
pid to the argument varibale and in eval when waitfg is called it will
check if the pid is a fg process and then wait for it accordingly.